+++ static_precompiler/tests/test_coffeescript.py
@@ -24,6 +24,21 @@ def test_compile_file(monkeypatch, tmpdir):
         assert clean_javascript(compiled.read()) == """(function() {\n  console.log("Hello, World!");\n}).call(this);"""
 
 
+def test_sourcemap(monkeypatch, tmpdir):
+
+    monkeypatch.setattr("static_precompiler.settings.ROOT", tmpdir.strpath)
+
+    compiler = compilers.CoffeeScript(sourcemap_enabled=False)
+    compiler.compile_file("scripts/test.coffee") == "COMPILED/scripts/test.js"
+    full_output_path = compiler.get_full_output_path("scripts/test.coffee")
+    assert not os.path.exists(os.path.splitext(full_output_path)[0] + ".map")
+
+    compiler = compilers.CoffeeScript(sourcemap_enabled=True)
+    compiler.compile_file("scripts/test.coffee") == "COMPILED/scripts/test.js"
+    full_output_path = compiler.get_full_output_path("scripts/test.coffee")
+    assert os.path.exists(os.path.splitext(full_output_path)[0] + ".map")
+
+
 def test_compile_source():
     compiler = compilers.CoffeeScript()